incorporated Notes into noteboard
