+++ client/Routes.jsx
@@ -11,7 +11,8 @@ import SocketIOContainer from './containers/SocketIOContainer';
 
 //action-creators
 import {getBoard, getAllBoards} from './actions/board';
-import {getNotes} from './actions/noteboard';
+// import {getNotes} from './actions/noteboard';
+import {getAllNotes} from './actions/note';
 
 //components
 import Index from './components/Index';
@@ -24,7 +25,7 @@ export const io =  socketClient;
 function onBoardEnter(nextRouterState) {
   const boardId = nextRouterState.params.boardId;
   store.dispatch(getBoard(boardId));
-  store.dispatch(getNotes());
+  store.dispatch(getAllNotes({boardId}));
 }
 
 function onMyBoardEnter(nextRouterState) {
