+++ client/reducers/noteReducer.js
@@ -1,8 +1,9 @@
-import {RECEIVE_NOTE, RECEIVE_NOTES, SELECT_NOTE} from '../constants';
+import {RECEIVE_NOTE, RECEIVE_NOTES, SELECT_NOTE, MOVE_NOTE } from '../constants';
 
 const initState = {
   all     : [],
-  selected: null
+  selected: null,
+
 };
 
 export default function noteReducer(state = initState, action) {
@@ -19,6 +20,14 @@ export default function noteReducer(state = initState, action) {
   case SELECT_NOTE:
     nextState.selected = nextState.all.find(note => note.id === action.payload.id);
     break;
+
+  case MOVE_NOTE:
+    const keys = Object.keys(action.notes);
+    keys.forEach((key) => {
+      nextState.all[key] = {left: action.notes[key].left, top: action.notes[key].top };
+    });
+    break;
+
 
   default:
     return state;
